fix(#146): 修复 PR #150 review 排查的 Blocker 与 Major 问题
Blocker（编译/签名）:
- B3 BasicProfileAnalyzer.filterIqr: Long 在 Double range 内比较类型不匹配，转 toDouble()
- B4 FeedbackController: shouldApply(scenarioId,sessionId)/selectDriven 改为 suspend（内部调 suspend getFeedbackGrayRatio）
- B5 FeedbackAgentPromptBuilder: 补 sanitize import；用户输入用 <<<USER_INPUT>>> 边界标记包裹，system prompt 增注入约束
- B6/B7 DevOptionsScreen: 新增 onNavigateToProfileChat + ProfileInspectViewModel 参数，落地"我的画像"展示与回滚入口

core-scheduler Major:
- M1 UserProfileWorker.computeFingerprint: 剔除 latestVersion.id（输出字段），避免缓存命中短路永远失效
- M2 LLM 调用前显式 getDefaultProvider() 判空，用 getClient(provider) 替代 getDefaultClient()
- M4 promptHash 加入 PROMPT_TEMPLATE_VERSION 常量，模板变更可识别

core-profiling feedback Major:
- M-反馈1 ProfileAdjuster: 以 mutated 标志决定缓存失效，覆盖 EnableScenario 软删无新记录场景
- M-反馈2 UserProfileOverrideDao: 新增 @Transaction markSupersededAndInsert，软删+插入原子化
- M-反馈3 ProfileVersionStore: 注入 AppDatabase，applyRollback 多步写包 withTransaction
- M-反馈4 FeedbackAgent: inFlight 预留计数，关闭 tryAcquireRate 的 check-then-act 竞争窗口

core-llm Major:
- UserProfilePromptBuilder: narrative 增安全约束（禁止虚假身份/暴力/仇恨/色情）

app 埋点 Major:
- MainActivity: 9 路由 SCREEN_ENTER/LEAVE + Tab TAB_SWITCH 埋点

依赖:
- app: += kotlinx-serialization-json（埋点 extra 需 JsonElement）
- core-profiling: += room-runtime/room-ktx（withTransaction 扩展）

B1 经核实为误判（entity 有默认值，迁移一致）；B2 为历史遗留（6.json 亦缺），非 #146 引入。

+++ core-profiling/src/main/java/com/trae/social/core/profiling/analysis/BasicProfileAnalyzer.kt
@@ -309,7 +309,8 @@ object BasicProfileAnalyzer {
         val iqr = q3 - q1
         val lower = q1 - 1.5 * iqr
         val upper = q3 + 1.5 * iqr
-        val filtered = sorted.filter { it in lower..upper }
+        // B3 修复：it 为 Long，lower/upper 为 Double，需转 Double 后再比较，否则类型不匹配编译失败
+        val filtered = sorted.filter { it.toDouble() in lower..upper }
         return filtered to (sorted.size - filtered.size)
     }
